requirements.txt
requests
python-dotenv
flask
requests
beautifulsoup4
pandas
flask_cors
pymongo